backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepgram-sdk>=4.8.1",
    "fastapi>=0.117.1",
    "google-genai>=1.40.0",
    "httpx>=0.28.1",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.37.0",
]
